+++ plugins/partiql-mockdb/src/main/kotlin/org/partiql/plugins/mockdb/TrimLead.kt
@@ -4,16 +4,18 @@ package org.partiql.plugins.mockdb
 import org.partiql.spi.connector.ConnectorSession
 import org.partiql.spi.function.PartiQLFunction
 import org.partiql.spi.function.PartiQLFunctionExperimental
-import org.partiql.value.PartiQLValueType
 import org.partiql.types.function.FunctionParameter
 import org.partiql.types.function.FunctionSignature
 import org.partiql.value.PartiQLValue
 import org.partiql.value.PartiQLValueExperimental
+import org.partiql.value.PartiQLValueType
 import org.partiql.value.StringValue
 import org.partiql.value.stringValue
 
 @OptIn(PartiQLFunctionExperimental::class)
 object TrimLead : PartiQLFunction {
+
+    @OptIn(PartiQLValueExperimental::class)
     override val signature = FunctionSignature(
         name = "trim_lead",
         returns = PartiQLValueType.STRING,